requirement.txt
flask
streamlit
pandas
numpy
scikit-learn
plotly
openai
requests
gdown
gunicorn
taipy
google-auth
google-api-python-client
Dash
matplotlib
seaborn
scikit-learn
panel
mlflow
transformers 
torch